// src/components/cdk/cdk-forseti-preview/models/cdk-forseti-preview.model.ts
export interface ICDKForsetiPreviewLiterals {
	contractsLeyend: string;
	titleStatus: string;
	subMessage: string;
	errorTilte: string;
	errorDesc: string;
}

export interface ICDKForsetiPreviewData {
	totalContracts: string | number;
	totalPending: string | number;
	ilustration: string;
	loadingImgSrc: string;
}